0024355: Compiler Warning level 4 for MFC samples
Some warnings were removed from MFC samples
All warnings of sample projects with warning level 4 were fixed on VC9
Warnings level increased to -W4 in projects of MFC samples; sample ReadMe files added to documentation
Changes in Viewer2dDoc.cpp and OcafDoc.cpp were restored.
Minor corrections in documentation

+++ samples/mfc/standard/05_ImportExport/src/FSD_Archive.cxx
@@ -52,7 +52,7 @@ Storage_Error FSD_Archive::IsGoodFileType(const TCollection_AsciiString&
 
   if (s == Storage_VSOk) {
     TCollection_AsciiString l;
-    Standard_Integer        len = strlen(FSD_Archive::MagicNumber());
+    Standard_Integer len = (int)strlen(FSD_Archive::MagicNumber());
 
     f.ReadChar(l,len);
 
@@ -851,7 +851,7 @@ Storage_Error FSD_Archive::BeginReadInfoSection()
   Storage_Error s;
 #ifdef WNT
   TCollection_AsciiString l;
-  Standard_Integer        len = strlen(FSD_Archive::MagicNumber());
+  Standard_Integer len = (int)strlen(FSD_Archive::MagicNumber());
 
   // first character is length of the magic number
   ReadChar(l,1);
